Fix: ViewExtension 제거
