+++ Timery/Modules/UI/Home/Sources/Home/HomeVC.swift
@@ -6,9 +6,7 @@
 //
 
 import UIKit
-import SnapKit
-import Then
-import ViewExtension
+import UIFusionKit
 
 public final class HomeVC: UITabBarController {
