+++ Timery/Modules/UI/TestDetail/Package.swift
@@ -15,7 +15,7 @@ let package = Package(
             targets: ["TestDetail"]),
     ],
     dependencies: [
-        .package(path: "../ViewExtension")
+        .package(path: "../UIFusionKit")
     ],
     targets: [
         // Targets are the basic building blocks of a package, defining a module or a test suite.
@@ -23,7 +23,7 @@ let package = Package(
         .target(
             name: "TestDetail",
             dependencies: [
-                "ViewExtension"
+                "UIFusionKit"
             ]
         ),
         .testTarget(
