+++ Timery/Modules/UI/UIFusionKit/Package.resolved
@@ -1,5 +1,5 @@
 {
-  "originHash" : "e34045e3ffaed0a63ef68789add69cf31b1b2721d0019ebd406b72268e7a6002",
+  "originHash" : "9cb3db034de9be1f8ba8eb2e3bb6efaf1aae6f2ad22858f08111325784638677",
   "pins" : [
     {
       "identity" : "snapkit",
