+++ Timery/Modules/UI/UIFusionKit/Tests/UIFusionKitTests/UIFusionKitTests.swift
@@ -1,0 +1,17 @@
+//
+//  UIFusionKitViewController.swift
+//  Timery
+//
+//  Created by 정준영 on 2024/06/16
+//
+
+import XCTest
+@testable import UIFusionKit
+
+class UIFusionKitTests: XCTestCase {
+    func testExample() {
+        // Example test case
+        let viewController = UIFusionKitViewController()
+        XCTAssertNotNil(viewController.view)
+    }
+}
