--- Timery/Modules/UI/ViewExtension/Tests/ViewExtensionTests/ViewExtensionTests.swift
@@ -1,12 +1,0 @@
-import XCTest
-@testable import ViewExtension
-
-final class ViewExtensionTests: XCTestCase {
-    func testExample() throws {
-        // XCTest Documentation
-        // https://developer.apple.com/documentation/xctest
-
-        // Defining Test Cases and Test Methods
-        // https://developer.apple.com/documentation/xctest/defining_test_cases_and_test_methods
-    }
-}
