+++ timery/Project.swift
@@ -28,9 +28,9 @@ let deploymentTarget: DeploymentTargets = .iOS("17.0")
 let project = Project(
     name: "Timery",
     packages: [
-        .local(path: .relativeToRoot("Modules/UI/ViewExtension")),
         .local(path: .relativeToRoot("Modules/UI/Home")),
         .local(path: .relativeToRoot("Modules/UI/TestDetail")),
+        .local(path: .relativeToRoot("Modules/UI/UIFusionKit"))
     ],
     targets: [
         .target(
@@ -43,9 +43,9 @@ let project = Project(
             sources: ["App/Sources/**"],
             resources: ["App/Resources/**"],
             dependencies: [
-                .package(product: "ViewExtension"),
                 .package(product: "Home"),
                 .package(product: "TestDetail"),
+                .package(product: "UIFusionKit"),
             ]
         ),
         .target(
